python/requirements.txt
altgraph==0.17.4
bidict==0.22.1
blinker==1.7.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
Flask==3.0.2
Flask-SocketIO==5.3.6
gevent==23.9.1
gevent-websocket==0.10.1
greenlet==3.0.3
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==23.2
pefile==2023.2.7
pycparser==2.21
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.0
python-engineio==4.9.0
python-socketio==5.11.1
pywin32-ctypes==0.2.2
simple-websocket==1.0.0
Werkzeug==3.0.1
wsproto==1.2.0
zope.event==5.0
zope.interface==6.1